day 10 part 2 solved
How's this for horrible: I got 2/3 tests to pass and I noticed that the
failing one was identifying outside instead of inside. So I ran my code
against the real data, knowing that I could either take that answer or
"negate" it. Turns out running it as-is worked. So I feel horrible, but I
got the right answer.